TACS-31 refactor: remove in memory repository
