+++ auth/src/auth/adapters/repository.py
@@ -2,7 +2,7 @@
 This module abstracts the Database layer with a Repository pattern.
 """
 import abc
-from typing import Generic, TypeVar
+from typing import TypeVar
 
 from auth.domain.models import BaseEntity, User
 
@@ -91,37 +91,3 @@ class UserRepository(Repository, abc.ABC):
 
         """
         raise NotImplementedError
-
-
-class InMemoryRepository(Repository, Generic[T]):
-    """
-    An in-memory repository implementation.
-    """
-
-    def __init__(self, data: dict[str, T] | None = None):
-        self._data: dict[str, T] = data or dict()
-
-    def find_all(self) -> list[T]:
-        return list(self._data.values())
-
-    def find_by(self, **kwargs) -> T | None:
-        properties = kwargs.keys()
-        return next(
-            (entity for entity in self._data.values()
-             if all(getattr(entity, p) == kwargs[p] for p in properties)),
-            None)
-
-    def save(self, entity: T) -> None:
-        self._data[entity.id] = entity
-
-    def delete(self, entity: T) -> None:
-        del self._data[entity.id]
-
-
-class InMemoryUserRepository(InMemoryRepository[User], UserRepository):
-    """
-    An in-memory user repository implementation.
-    """
-
-    def find_by_username(self, username: str) -> User | None:
-        return self.find_by(username=username)
